+++ barapost/src/lingering_https_get_request.py
@@ -35,12 +35,6 @@ def lingering_https_get_request(server, url, logfile_path, request_for=None, acc
             conn.request("GET", url) # ask for if there areresults
             response = conn.getresponse() # get the resonse
 
-            # # 
-            # if response.code == 301 and request_for.startswith("link to identical genbank sequence"):
-            #     return lingering_https_get_request(server, response.getheader("Location")+"?report=accnlist&log$=seqview&format=text",
-            #         logfile_path, request_for, acc)
-            # # end if
-
             if response.code != 200:
                 printl(logfile_path, "Request failed with status code {}: {}".format(response.code, response.reason))
                 platf_depend_exit(1)